test pasing

+++ arcus/exc.py
@@ -82,7 +82,7 @@ class RecurrentPayments(UnprocessableEntity):
 
 
 class DuplicatedPayment(UnprocessableEntity):
-    def __init__(self, code: str, amount: int):
+    def __init__(self, code: str, amount: float):
         message = f'Duplicated payment for {amount}'
         super().__init__(
             code, message
@@ -90,7 +90,7 @@ class DuplicatedPayment(UnprocessableEntity):
 
 
 class IncompleteAmount(UnprocessableEntity):
-    def __init__(self, code: str, amount: int):
+    def __init__(self, code: str, amount: float):
         message = (
             f'Incomplete payment amount of {amount}, must pay full balance')
         super().__init__(
